Add yarn.lock file with package versions
The yarn.lock file was added to the repository to lock the versions of the dependencies. This ensures that all developers working on the project use the same versions of the packages. When a developer installs a package using Yarn, the exact version of the dependencies gets saved in yarn.lock, and this file should be committed to the version control system.